[feat][lanie] removed the dashboard/case/ page and fixed the error in browse case component
